Create Institution Endpoint (SOFTWARE-5443)
Move the create_accepted_response function
