+++ src/webapp/common.py
@@ -15,7 +15,6 @@ import xmltodict
 import yaml
 import csv
 from io import StringIO
-from flask import Response
 
 try:
     from yaml import CSafeLoader as SafeLoader
@@ -55,26 +54,6 @@ class Filters(object):
 
     def populate_voown_name(self, vo_id_to_name: Dict):
         self.voown_name = [vo_id_to_name.get(i, "") for i in self.voown_id]
-
-
-def create_accepted_response(data: list, headers, default=None) -> Response:
-    """Provides CSV or JSON options for list of list(string)"""
-
-    if not default:
-        default = "application/json"
-
-    accepted_response_builders = {
-        "text/csv": lambda: Response(to_csv(data), mimetype="text/csv"),
-        "application/json": lambda: Response(to_json_bytes(data), mimetype="application/json"),
-    }
-
-    requested_types = set(headers.get('Accept', 'default').replace(' ', '').split(","))
-    accepted_and_requested = set(accepted_response_builders.keys()).intersection(requested_types)
-
-    if accepted_and_requested:
-        return accepted_response_builders[accepted_and_requested.pop()]()
-    else:
-        return accepted_response_builders[default]()
 
 
 def to_csv(data: list) -> str:
